requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.0.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
